fix stop script

+++ ubuntu/scripts/stop_pl_server.py
@@ -42,4 +42,5 @@ try:
 except ConnectionRefusedError:
     pass
     
-os.remove("/home/xpp/pynq/bitstream/.log")
+if os.path.isfile("/home/xpp/pynq/bitstream/.log"):
+    os.remove("/home/xpp/pynq/bitstream/.log")
